apps/api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
supabase>=2.11.0
pyjwt>=2.10.0
stripe>=11.0.0
python-dotenv>=1.0.0
httpx>=0.28.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
